+++ packages/dynamic_morphing_icons/lib/dynamic_morphing_icons.dart
@@ -14,3 +14,4 @@ export 'src/morph_icon.dart';
 export 'src/morph_icons.dart';
 export 'src/morph_painter.dart';
 export 'src/morphing_icon.dart';
+export 'src/morphing_icon_theme.dart';